refactor: engine.py 关注点分离，拆分为 filters/matcher/scanner/rule_generator 四个模块
将 1309 行的 engine.py 按职责拆分为 5 个文件：
- core/filters.py (135行): VulnerabilityFilter 文件前置过滤
- core/matcher.py (409行): VulnerabilityMatcher(=Core) 漏洞判定引擎，按语言分派，提取 _parse_ast_result() 消除 PHP/JS 重复代码
- core/scanner.py (548行): Running/scan/score2level/SingleRule 扫描调度层
- core/rule_generator.py (205行): NewCore/init_match_rule/auto_parse_match 规则生成
- core/engine.py (24行): 薄兼容层 re-export，所有旧 import 路径保持有效

测试结果：90 passed / 1 failed（与基线一致，失败为已有的 test_phpunserializechain_complex_chain）
